sort: print warning when fork() failed for --compress-program
sort already have one when it trying to create decompressor, it is
obvious why it is really required in this case, since sort will read
compressed data as plain otherwise.
But sometimes it is really usefull to know whether sort failed to create
compressor or not, since some users may rely on available free space and
compressor.

* src/sort.c (create_temp_file): Add a warning when creating of
compressor failed.