sort: use pre-fork at start for compress-prog, to avoid copying VM
sort uses fork() to run compress-prog (like gzip) however fork() could
fail with default overcommit (on linux) when sort uses buffer-size=90%
for example, because fork() needs to copy the vm for new process and
will validate that we have enough memory for this (i.e. after COW
will triggered on every page), so to fix this fork() at the beginning
of the program when we don't have huge buffer allocated and use pipe
to interact (i.e. run/wait compress-prog) with it.

P.S. this is just WIP patch, so yes it need cleanup and some fixes,
just to share the progress of it.

Link: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=203115